third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/iterable-expected.txt

This is a testharness.js-based test.
PASS Iterating over an empty StylePropertyMap gives a zero-length array
FAIL StylePropertyMap iterates properties in correct order assert_array_equals: property 0, expected "color" but got "--A"
FAIL StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue assert_equals: expected "width" but got "--C"
FAIL StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue assert_equals: expected "transition-duration" but got "width"
FAIL StylePropertyMap iterator returns custom properties with the correct CSSStyleValue assert_equals: expected "--A" but got "transition-duration"
Harness: the test ran to completion.
